fix(bootstrap): move dashboard admin RPC into dashboard/ subfolder
Satisfies folder file-count gate (max 15) on bootstrap/factory.
